Connection Strength Bars tweak
So whenever the connection strength changed, the size of the UserBar
View changed and the whole interface would shift over a bit. This
started to annoy me so here's a try at fixing it, using
Visibility.Hidden instead of Visibility.Collapsed.

As before, let me know if anything different / more should be done! Be
picky if you feel like it